Show update progress
Wanted to host on google drive but disappointingly and unfortunately
out of the 3 GET/POST/HEAD for HEAD it's returning 405 ie, not supported
error, thus providing no way of knowing Content-Length ie, the file size!!

So must host it now on github itself.

So added gadd-bin to add bin/compiled files which can be pused on to
git but won't be tracked for changes. Sad.